requirements.txt
beautifulsoup4==4.10.0
requests==2.28.0
ipykernel==6.20.1
pandas
numpy
scikit-learn
matplotlib
seaborn
flask
python-dotenv
# -e .